Move 'ZipFileUtil' to sub-project java-extras
